+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # v0.2.2 (git develop)
 
+- IdFactory is now configurable per class. Some classes may use IdFactory::AtomicIncrementation,
+  others might use IdFactory::Timestamp. By default we are using AtomicIncrementation.
+  It is configurable via Person.id_factory = MassiveRecord::ORM::IdFactory::AtomicIncrementation
+- Added an IdFactory which uses Time as its generator, instead of atomic incrementation of a value.
 - ORM records now responds to raw_data(). it will give you a nested hash corresponding to column families
   and cells with RawData objects as values. These RawData objects contains the raw value and meta data like
   timestamp on the cell from Hbase.
@@ -47,7 +51,7 @@
 # v0.2.1 (git master)
 
 - Models without any default_id will now by default get an id via next_id(). You can turn it off
-  via the setting auto_increment_id on ORM::Base or on the Model class itself.
+  via the setting set_id_from_factory_before_create on ORM::Base or on the Model class itself.
 - record.reload now resets relations.
 - If you have a persisted record and you set one attribute to nil that attribute will be
   deleted from HBase to represent the nil-value. The fact that the schema of that record